pyproject.toml
[project]
name = "lpo-measure"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.75.8",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "orjson>=3.11.2",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pyright>=1.1.403",
]
